feat: add relation between invoices and user

+++ src/invoice/invoice.service.ts
@@ -7,18 +7,35 @@ import { CreateInvoiceDto, DraftInvoiceDto } from './dto/'
 import { InvoiceEntity } from './invoice.entity'
 import { addDays } from 'utils'
 import { IInvoiceStatus, IInvoicesQuery, IInvoicesResponse } from './types'
+import { UserEntity } from 'auth/auth.entity'
 
 @Injectable()
 export class InvoiceService {
   constructor(@InjectRepository(InvoiceEntity) private readonly invoiceRepository: Repository<InvoiceEntity>) {}
 
   /**
-   * It takes an object of type IInvoicesQuery, and returns an object of type IInvoicesResponse
+   * It returns a list of invoices for a given user, with optional filtering by status and pagination.
+   *
+   * The function takes two arguments:
+   *
+   * query: IInvoicesQuery - an object containing the query parameters
+   * currentUserId: number - the ID of the user whose invoices we want to retrieve
+   * The function returns an object of type IInvoicesResponse, which contains the following properties:
+   *
+   * data: Invoice[] - an array of Invoice objects
+   * count: number - the total number of invoices for the given user, ignoring pagination
+   * The query object has the following properties:
+   *
+   * status: string[] - an array of invoice statuses to filter by
+   * limit: number - the maximum number of invoices to return
+   * offset: number - the number of invoices to skip before returning results
+   * The function uses the
    * @param {IInvoicesQuery} query - IInvoicesQuery
-   * @returns An array of invoices and the total count of invoices.
+   * @param {number} currentUserId - number - this is the user id of the user who is logged in
+   * @returns An array of Invoice entities and the total count of Invoice entities.
    */
-  async getInvoices(query: IInvoicesQuery): Promise<IInvoicesResponse> {
-    const queryBuilder = this.invoiceRepository.createQueryBuilder('invoices')
+  async getInvoices(query: IInvoicesQuery, currentUserId: number): Promise<IInvoicesResponse> {
+    const queryBuilder = this.invoiceRepository.createQueryBuilder('invoices').where('invoices.ownerId = :id', { id: currentUserId })
 
     if (query.status?.length > 0) {
       queryBuilder.andWhere('invoices.status IN (:...status)', { status: query.status })
@@ -41,29 +58,34 @@ export class InvoiceService {
   }
 
   /**
-   * This function creates a new invoice with the status of pending.
+   * "Create a new invoice with the given data, set the status to pending, and return the invoice."
    * @param {CreateInvoiceDto} createInvoiceDto - CreateInvoiceDto
+   * @param {UserEntity} currentUser - UserEntity
    * @returns The return type is InvoiceEntity.
    */
-  async createNewInvoice(createInvoiceDto: CreateInvoiceDto): Promise<InvoiceEntity> {
-    return await this.createInvoice(createInvoiceDto, IInvoiceStatus.PENDING)
+  async createNewInvoice(createInvoiceDto: CreateInvoiceDto, currentUser: UserEntity): Promise<InvoiceEntity> {
+    return await this.createInvoice(createInvoiceDto, IInvoiceStatus.PENDING, currentUser)
   }
 
   /**
-   * This function takes a DraftInvoiceDto and returns an InvoiceEntity
+   * This function takes a DraftInvoiceDto and a UserEntity and returns a Promise of an InvoiceEntity.
    * @param {DraftInvoiceDto} draftInvoiceDto - DraftInvoiceDto
-   * @returns The return type is a Promise of an InvoiceEntity.
+   * @param {UserEntity} currentUser - UserEntity
+   * @returns The return type is InvoiceEntity.
    */
-  async addDraftInvoice(draftInvoiceDto: DraftInvoiceDto): Promise<InvoiceEntity> {
-    return await this.createInvoice(draftInvoiceDto, IInvoiceStatus.DRAFT)
+  async addDraftInvoice(draftInvoiceDto: DraftInvoiceDto, currentUser: UserEntity): Promise<InvoiceEntity> {
+    return await this.createInvoice(draftInvoiceDto, IInvoiceStatus.DRAFT, currentUser)
   }
 
   /**
-   * It finds an invoice by its id, and if it exists, it updates it with the new data
-   * @param {InvoiceEntity} invoiceToUpdate - InvoiceEntity
+   * "If the invoice exists, and the current user is the owner of the invoice, then update the
+   * invoice."
+   * @param {InvoiceEntity} invoiceToUpdate - InvoiceEntity - this is the invoice object that is passed
+   * in from the frontend.
+   * @param {number} currentUserId - number - this is the id of the user who is currently logged in.
    * @returns The updated invoice
    */
-  async editInvoice(invoiceToUpdate: InvoiceEntity): Promise<InvoiceEntity> {
+  async editInvoice(invoiceToUpdate: InvoiceEntity, currentUserId: number): Promise<InvoiceEntity> {
     const invoice = await this.invoiceRepository.findOne({
       where: {
         id: invoiceToUpdate.id,
@@ -74,17 +96,22 @@ export class InvoiceService {
       throw new HttpException('Invoice is not exist', HttpStatus.NOT_FOUND)
     }
 
+    if (invoice.owner.id !== currentUserId) {
+      throw new HttpException('You are not an owner of invoice', HttpStatus.FORBIDDEN)
+    }
+
     return await this.invoiceRepository.save({
       ...invoiceToUpdate,
     })
   }
 
   /**
-   * It finds an invoice by its ID, checks if it exists, and if it does, it changes its status to PAID
+   * Mark an invoice as paid by its id and the id of the current user.
    * @param {number} invoiceId - number - the id of the invoice to be marked as paid
-   * @returns InvoiceEntity
+   * @param {number} currentUserId - number - this is the id of the user who is currently logged in.
+   * @returns The invoice entity with the updated status.
    */
-  async markInvoiceAsPaid(invoiceId: number): Promise<InvoiceEntity> {
+  async markInvoiceAsPaid(invoiceId: number, currentUserId: number): Promise<InvoiceEntity> {
     const invoice = await this.invoiceRepository.findOne({
       where: {
         id: invoiceId,
@@ -93,6 +120,10 @@ export class InvoiceService {
 
     if (!invoice) {
       throw new HttpException('Invoice is not exist', HttpStatus.NOT_FOUND)
+    }
+
+    if (invoice.owner.id !== currentUserId) {
+      throw new HttpException('You are not an owner of invoice', HttpStatus.FORBIDDEN)
     }
 
     invoice.status = IInvoiceStatus.PAID
@@ -101,13 +132,20 @@ export class InvoiceService {
   }
 
   /**
-   * It creates a new invoice, assigns it a random order ID, calculates the payment due date, and saves
-   * it to the database
+   * "Create a new invoice, assign it a random orderId, assign it a paymentDue date if the user has
+   * provided a createdAt and paymentTerms, assign it a status, assign it an owner, and then assign it
+   * the rest of the properties from the createInvoiceDto."
+   * </code>
    * @param {DraftInvoiceDto | CreateInvoiceDto} createInvoiceDto - DraftInvoiceDto | CreateInvoiceDto
    * @param {IInvoiceStatus} status - IInvoiceStatus
-   * @returns The new invoice is being returned.
+   * @param {UserEntity} currentUser - UserEntity
+   * @returns The newInvoice object is being returned.
    */
-  async createInvoice(createInvoiceDto: DraftInvoiceDto | CreateInvoiceDto, status: IInvoiceStatus): Promise<InvoiceEntity> {
+  async createInvoice(
+    createInvoiceDto: DraftInvoiceDto | CreateInvoiceDto,
+    status: IInvoiceStatus,
+    currentUser: UserEntity
+  ): Promise<InvoiceEntity> {
     const newInvoice = new InvoiceEntity()
     newInvoice.orderId = new RandExp(/([A-Z]{2}[0-9]{4})/).gen()
 
@@ -116,17 +154,19 @@ export class InvoiceService {
     }
 
     newInvoice.status = status
+    newInvoice.owner = currentUser
 
     Object.assign(newInvoice, createInvoiceDto)
     return await this.invoiceRepository.save(newInvoice)
   }
 
   /**
-   * It deletes an invoice by its id.
+   * Delete an invoice by id, if the invoice exists and the current user is the owner of the invoice.
    * @param {number} invoiceId - number - the id of the invoice to be deleted
-   * @returns The return type is a Promise&lt;DeleteResult&gt;.
+   * @param {number} currentUserId - number - this is the id of the user who is currently logged in.
+   * @returns The return value is the number of rows affected by the delete query.
    */
-  async deleteInvoice(invoiceId: number) {
+  async deleteInvoice(invoiceId: number, currentUserId: number) {
     const invoice = await this.invoiceRepository.findOne({
       where: {
         id: invoiceId,
@@ -135,6 +175,10 @@ export class InvoiceService {
 
     if (!invoice) {
       throw new HttpException('Invoice is not exist', HttpStatus.NOT_FOUND)
+    }
+
+    if (invoice.owner.id !== currentUserId) {
+      throw new HttpException('You are not an owner of invoice', HttpStatus.FORBIDDEN)
     }
 
     return await this.invoiceRepository.delete({ id: invoiceId })
